# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SahiPath API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResult {
  error: string;
}

export interface OkResult {
  ok: boolean;
}

export interface UserProfile {
  language?: string;
  firstName?: string;
  lastName?: string;
  /** @nullable */
  age?: number | null;
  email?: string;
  location?: string;
  educationLevel?: string;
  currentRole?: string;
  /** @nullable */
  yearsOfExperience?: number | null;
  skills?: string[];
  careerInterests?: string[];
  currentGoals?: string;
  challenges?: string;
  /** @nullable */
  availableHoursPerWeek?: number | null;
  preferredLearningStyle?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  profile?: UserProfile;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserRecord {
  id: string;
  email: string;
  profile?: UserProfile;
}

export interface AuthResult {
  user: UserRecord;
}

export interface ProfileResult {
  profile: UserProfile;
}

export type MentorChatInputMode = typeof MentorChatInputMode[keyof typeof MentorChatInputMode];


export const MentorChatInputMode = {
  text: 'text',
  voice: 'voice',
} as const;

export interface MentorChatInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  message: string;
  mode?: MentorChatInputMode;
  profile?: UserProfile;
}

export interface MentorChatResult {
  reply: string;
}

export interface TestInput {
  name: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  date?: string;
  notes?: string;
}

export interface TestRecord {
  id: string;
  name: string;
  score: number;
  date: string;
  notes?: string;
}

export interface TestsResult {
  tests: TestRecord[];
}

export interface TestRecordResult {
  test: TestRecord;
}

export type MediaInputType = typeof MediaInputType[keyof typeof MediaInputType];


export const MediaInputType = {
  image: 'image',
  podcast: 'podcast',
  video: 'video',
  ppt: 'ppt',
} as const;

export interface MediaInput {
  type: MediaInputType;
  /**
     * @minLength 1
     * @maxLength 5000
     */
  prompt: string;
}

export interface MediaResult {
  id: string;
  type: string;
  url?: string;
  note?: string;
}