# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SahiPath API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: mentor
    description: AI Career Mentor chat
  - name: profile
    description: User profile
  - name: tests
    description: Test records
  - name: media
    description: AI Media generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /mentor/chat:
    post:
      operationId: mentorChat
      tags: [mentor]
      summary: Send message to AI career mentor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorChatInput"
      responses:
        "200":
          description: AI reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorChatResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "503":
          description: Not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
    post:
      operationId: saveProfile
      tags: [profile]
      summary: Save user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfile"
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /tests:
    get:
      operationId: getTests
      tags: [tests]
      summary: Get test records
      responses:
        "200":
          description: Tests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestsResult"
    post:
      operationId: recordTest
      tags: [tests]
      summary: Record a test result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestInput"
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestRecordResult"

  /media:
    post:
      operationId: generateMedia
      tags: [media]
      summary: Generate AI media (image, podcast, video)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInput"
      responses:
        "200":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "503":
          description: Not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResult:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    OkResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        profile:
          $ref: "#/components/schemas/UserProfile"

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    UserRecord:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
        email:
          type: string
        profile:
          $ref: "#/components/schemas/UserProfile"

    AuthResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/UserRecord"

    UserProfile:
      type: object
      properties:
        language:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        age:
          type: ["integer", "null"]
        email:
          type: string
        location:
          type: string
        educationLevel:
          type: string
        currentRole:
          type: string
        yearsOfExperience:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        careerInterests:
          type: array
          items:
            type: string
        currentGoals:
          type: string
        challenges:
          type: string
        availableHoursPerWeek:
          type: ["integer", "null"]
        preferredLearningStyle:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    ProfileResult:
      type: object
      required: [profile]
      properties:
        profile:
          $ref: "#/components/schemas/UserProfile"

    MentorChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4000
        mode:
          type: string
          enum: [text, voice]
          default: text
        profile:
          $ref: "#/components/schemas/UserProfile"

    MentorChatResult:
      type: object
      required: [reply]
      properties:
        reply:
          type: string

    TestInput:
      type: object
      required: [name, score]
      properties:
        name:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
        date:
          type: string
        notes:
          type: string

    TestRecord:
      type: object
      required: [id, name, score, date]
      properties:
        id:
          type: string
        name:
          type: string
        score:
          type: number
        date:
          type: string
        notes:
          type: string

    TestsResult:
      type: object
      required: [tests]
      properties:
        tests:
          type: array
          items:
            $ref: "#/components/schemas/TestRecord"

    TestRecordResult:
      type: object
      required: [test]
      properties:
        test:
          $ref: "#/components/schemas/TestRecord"

    MediaInput:
      type: object
      required: [type, prompt]
      properties:
        type:
          type: string
          enum: [image, podcast, video, ppt]
        prompt:
          type: string
          minLength: 1
          maxLength: 5000

    MediaResult:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
        url:
          type: string
        note:
          type: string