# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SahiPath API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRecord } from './userRecord';

export interface AuthResult {
  user: UserRecord;
}