# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SahiPath API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  language?: string;
  firstName?: string;
  lastName?: string;
  /** @nullable */
  age?: number | null;
  email?: string;
  location?: string;
  educationLevel?: string;
  currentRole?: string;
  /** @nullable */
  yearsOfExperience?: number | null;
  skills?: string[];
  careerInterests?: string[];
  currentGoals?: string;
  challenges?: string;
  /** @nullable */
  availableHoursPerWeek?: number | null;
  preferredLearningStyle?: string;
  createdAt?: string;
  updatedAt?: string;
}